Update README.md
This is a music streaming platform where the user will neither get any ads while listening to the songs nor have to download the favourite song for listening it. And it also helps users to modify and create their own playlist of their own choice. It will work as same as any other music platform like  YouTube  music and  Spotify. it's interface will be more interactive and engaging with users. The platform will be prepared with the help of HTML, CSS and Java-script.